requirements.txt
detectlanguage==1.3.0
Django==2.2.4
factory-boy==2.12.0
